+++ backend/src/main/java/com/s12p31b204/backend/controller/NoteController.java
@@ -40,12 +40,11 @@ public class NoteController {
             @AuthenticationPrincipal CustomOAuth2User user,
             HttpServletRequest request) {
         try {
-//            if (authorizationService.checkWorkBookAuthorization(user.getUserId(), workBookId)) {
-//                return ApiResponse.success(noteService.getNoteTestPaperByWorkBookId(workBookId), "success", HttpStatus.OK, request.getRequestURI());
-//            } else {
-//                return ApiResponse.failure("권한이 없습니다.", HttpStatus.FORBIDDEN, request.getRequestURI());
-//            }
-            return ApiResponse.success(noteService.getNoteTestPaperByWorkBookId(workBookId), "success", HttpStatus.OK, request.getRequestURI());
+            if (authorizationService.checkWorkBookAuthorization(user.getUserId(), workBookId)) {
+                return ApiResponse.success(noteService.getNoteTestPaperByWorkBookId(workBookId), "success", HttpStatus.OK, request.getRequestURI());
+            } else {
+                return ApiResponse.failure("권한이 없습니다.", HttpStatus.FORBIDDEN, request.getRequestURI());
+            }
         } catch (Exception e) {
             return ApiResponse.failure(e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR, request.getRequestURI());
         }
@@ -57,12 +56,11 @@ public class NoteController {
             @AuthenticationPrincipal CustomOAuth2User user,
             HttpServletRequest request) {
         try {
-//            if (authorizationService.checkTestAuthorization(user.getUserId(), testId)) {
-//                return ApiResponse.success(noteService.getNoteTest(testId), "success", HttpStatus.OK, request.getRequestURI());
-//            } else {
-//                return ApiResponse.failure("권한이 없습니다.", HttpStatus.FORBIDDEN, request.getRequestURI());
-//            }
-            return ApiResponse.success(noteService.getNoteTest(testId), "success", HttpStatus.OK, request.getRequestURI());
+            if (authorizationService.checkTestAuthorization(user.getUserId(), testId)) {
+                return ApiResponse.success(noteService.getNoteTest(testId), "success", HttpStatus.OK, request.getRequestURI());
+            } else {
+                return ApiResponse.failure("권한이 없습니다.", HttpStatus.FORBIDDEN, request.getRequestURI());
+            }
         } catch (Exception e) {
             return ApiResponse.failure(e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR, request.getRequestURI());
         }
@@ -75,14 +73,12 @@ public class NoteController {
             @AuthenticationPrincipal CustomOAuth2User user,
             HttpServletRequest request) {
         try {
-//            if(authorizationService.checkTestAuthorization(user.getUserId(), testId)) {
-//                noteService.updateNoteMemo(testId, memoRequest);
-//                return ApiResponse.success(null, "success", HttpStatus.OK, request.getRequestURI());
-//            } else {
-//                return ApiResponse.failure("권한이 없습니다.", HttpStatus.FORBIDDEN, request.getRequestURI());
-//            }
-            noteService.updateNoteMemo(testId, memoRequest);
-            return ApiResponse.success(null, "success", HttpStatus.OK, request.getRequestURI());
+            if(authorizationService.checkTestAuthorization(user.getUserId(), testId)) {
+                noteService.updateNoteMemo(testId, memoRequest);
+                return ApiResponse.success(null, "success", HttpStatus.OK, request.getRequestURI());
+            } else {
+                return ApiResponse.failure("권한이 없습니다.", HttpStatus.FORBIDDEN, request.getRequestURI());
+            }
         } catch (Exception e) {
             return ApiResponse.failure(e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR, request.getRequestURI());
         }
@@ -94,14 +90,12 @@ public class NoteController {
             @AuthenticationPrincipal CustomOAuth2User user,
             HttpServletRequest request) {
         try {
-//            if(authorizationService.checkTestAuthorization(user.getUserId(), testId)) {
-//                noteService.removeNoteMemo(testId);
-//                return ApiResponse.success(null, "success", HttpStatus.OK, request.getRequestURI());
-//            } else {
-//                return ApiResponse.failure("권한이 없습니다.", HttpStatus.FORBIDDEN, request.getRequestURI());
-//            }
-            noteService.removeNoteMemo(testId);
-            return ApiResponse.success(null, "success", HttpStatus.OK, request.getRequestURI());
+            if(authorizationService.checkTestAuthorization(user.getUserId(), testId)) {
+                noteService.removeNoteMemo(testId);
+                return ApiResponse.success(null, "success", HttpStatus.OK, request.getRequestURI());
+            } else {
+                return ApiResponse.failure("권한이 없습니다.", HttpStatus.FORBIDDEN, request.getRequestURI());
+            }
         } catch (Exception e) {
             return ApiResponse.failure(e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR, request.getRequestURI());
         }
@@ -113,14 +107,12 @@ public class NoteController {
             @AuthenticationPrincipal CustomOAuth2User user,
             HttpServletRequest request) {
         try {
-//            if (authorizationService.checkTestPaperAuthorization(user.getUserId(), testPaperId)) {
-//                List<Long> testIds = noteService.getTestIdsByTestPaperId(testPaperId);
-//                return ApiResponse.success(testIds, "문제 ID 리스트 조회 성공", HttpStatus.OK, request.getRequestURI());
-//            } else {
-//                return ApiResponse.failure("권한이 없습니다.", HttpStatus.FORBIDDEN, request.getRequestURI());
-//            }
-            List<Long> testIds = noteService.getTestIdsByTestPaperId(testPaperId);
-            return ApiResponse.success(testIds, "문제 ID 리스트 조회 성공", HttpStatus.OK, request.getRequestURI());
+            if (authorizationService.checkTestPaperAuthorization(user.getUserId(), testPaperId)) {
+                List<Long> testIds = noteService.getTestIdsByTestPaperId(testPaperId);
+                return ApiResponse.success(testIds, "문제 ID 리스트 조회 성공", HttpStatus.OK, request.getRequestURI());
+            } else {
+                return ApiResponse.failure("권한이 없습니다.", HttpStatus.FORBIDDEN, request.getRequestURI());
+            }
         } catch (Exception e) {
             return ApiResponse.failure(e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR, request.getRequestURI());
         }
